agent/requirements.txt
google-adk>=1.28.1
firebase-admin>=6.5.0
python-dotenv>=1.0.1
